#95 [REFACTOR] 페이지 컴포넌트 이동 및 라우트 복수 표기
